+++ background.js
@@ -47,6 +47,11 @@ const configs = {
 
     'hoverSpaceEnabled': false,
 
+    'showPreviewIconOnHoverEnabled': false,
+    'dotSize': 16,
+    'dotRemoveDelay': 500,
+    'dotHoverDelay': 300,
+
     'clickModifiedKey': 'None',
     'previewModeDisabledUrls': [],
     'previewModeWindowType': 'popup',